+++ bt_servant_engine/adapters/user_state.py
@@ -1,18 +1,144 @@
-"""User state adapter implementing the TinyDB-backed port."""
+"""User state adapter implementing the TinyDB-backed port.
+
+Provides TinyDB database wiring and user data access helpers.
+"""
 
 from __future__ import annotations
 
+from pathlib import Path
+from typing import TYPE_CHECKING, Any, Dict, List, Optional, cast
+
+from tinydb import Query, TinyDB
+
+from bt_servant_engine.core.config import config
 from bt_servant_engine.core.ports import UserStatePort
-from db import (
-    get_user_agentic_strength,
-    get_user_chat_history,
-    get_user_response_language,
-    is_first_interaction,
-    set_first_interaction,
-    set_user_agentic_strength,
-    set_user_response_language,
-    update_user_chat_history,
-)
+
+# Provide a QueryLike alias for static checkers; at runtime use Any.
+if TYPE_CHECKING:  # pragma: no cover - typing only
+    from tinydb.queries import QueryLike  # type: ignore
+else:
+    QueryLike = Any  # type: ignore[misc,assignment]
+
+# NOTE: consider moving to configuration if needed.
+CHAT_HISTORY_MAX = 5
+VALID_AGENTIC_STRENGTH = {"normal", "low", "very_low"}
+
+# Resolve from settings dynamically to satisfy static analysis.
+DATA_DIR = Path(getattr(config, "DATA_DIR", Path("/data")))
+DATA_DIR.mkdir(parents=True, exist_ok=True)
+
+DB_PATH = DATA_DIR / "db.json"
+_db = TinyDB(str(DB_PATH))
+
+
+def get_user_db() -> TinyDB:
+    """Return the process-wide TinyDB instance for users."""
+    return _db
+
+
+def get_user_chat_history(user_id: str) -> List[Dict[str, str]]:
+    """Retrieve chat history for the given user_id."""
+    q = Query()
+    cond = cast(QueryLike, q.user_id == user_id)
+    raw = get_user_db().table("users").get(cond)
+    result = cast(Optional[Dict[str, Any]], raw)
+    history = result.get("history", []) if result else []
+    return cast(List[Dict[str, str]], history)
+
+
+def update_user_chat_history(user_id: str, query: str, response: str) -> None:
+    """Append a user/bot exchange to chat history (bounded)."""
+    q = Query()
+    users_table = get_user_db().table("users")
+    cond = cast(QueryLike, q.user_id == user_id)
+    existing_raw = users_table.get(cond)
+    existing = cast(Optional[Dict[str, Any]], existing_raw)
+    history = cast(List[Dict[str, str]], existing.get("history", []) if existing else [])
+
+    history.append({"user_message": query, "assistant_response": response})
+
+    history = history[-CHAT_HISTORY_MAX:]
+    users_table.upsert({"user_id": user_id, "history": history}, cond)
+
+
+def get_user_response_language(user_id: str) -> Optional[str]:
+    """Get the user's preferred response language, or None if not set."""
+    q = Query()
+    cond = cast(QueryLike, q.user_id == user_id)
+    raw = get_user_db().table("users").get(cond)
+    result = cast(Optional[Dict[str, Any]], raw)
+    lang = result.get("response_language") if result else None
+    return cast(Optional[str], lang)
+
+
+def set_user_response_language(user_id: str, language: str) -> None:
+    """Set the user's preferred response language."""
+    q = Query()
+    db = get_user_db().table("users")
+    cond = cast(QueryLike, q.user_id == user_id)
+    existing_raw = db.get(cond)
+    existing = cast(Optional[Dict[str, Any]], existing_raw)
+    updated: Dict[str, Any] = (
+        existing.copy() if isinstance(existing, dict) else {"user_id": user_id}
+    )
+
+    updated["response_language"] = language
+    db.upsert(updated, cond)
+
+
+def get_user_agentic_strength(user_id: str) -> Optional[str]:
+    """Get the user's preferred agentic strength, or None if not set."""
+    q = Query()
+    cond = cast(QueryLike, q.user_id == user_id)
+    raw = get_user_db().table("users").get(cond)
+    result = cast(Optional[Dict[str, Any]], raw)
+    strength = result.get("agentic_strength") if result else None
+    if isinstance(strength, str) and strength in VALID_AGENTIC_STRENGTH:
+        return strength
+    return None
+
+
+def set_user_agentic_strength(user_id: str, strength: str) -> None:
+    """Persist the user's preferred agentic strength level."""
+    normalized = strength.strip().lower()
+    if normalized not in VALID_AGENTIC_STRENGTH:
+        raise ValueError(f"invalid agentic strength: {strength}")
+
+    q = Query()
+    db = get_user_db().table("users")
+    cond = cast(QueryLike, q.user_id == user_id)
+    existing_raw = db.get(cond)
+    existing = cast(Optional[Dict[str, Any]], existing_raw)
+    updated: Dict[str, Any] = (
+        existing.copy() if isinstance(existing, dict) else {"user_id": user_id}
+    )
+    updated["agentic_strength"] = normalized
+    db.upsert(updated, cond)
+
+
+def set_first_interaction(user_id: str, is_first: bool) -> None:
+    """Set whether this is the user's first interaction."""
+    q = Query()
+    db = get_user_db().table("users")
+    cond = cast(QueryLike, q.user_id == user_id)
+    existing_raw = db.get(cond)
+    existing = cast(Optional[Dict[str, Any]], existing_raw)
+    updated: Dict[str, Any] = (
+        existing.copy() if isinstance(existing, dict) else {"user_id": user_id}
+    )
+
+    updated["first_interaction"] = is_first
+    db.upsert(updated, cond)
+
+
+def is_first_interaction(user_id: str) -> bool:
+    """Return True if this is the user's first interaction, otherwise False."""
+    q = Query()
+    cond = cast(QueryLike, q.user_id == user_id)
+    raw = get_user_db().table("users").get(cond)
+    result = cast(Optional[Dict[str, Any]], raw)
+    first = (result or {}).get("first_interaction", True)
+    return bool(first)
 
 
 class UserStateAdapter(UserStatePort):
@@ -43,4 +169,14 @@ class UserStateAdapter(UserStatePort):
         return is_first_interaction(user_id)
 
 
-__all__ = ["UserStateAdapter"]
+__all__ = [
+    "UserStateAdapter",
+    "get_user_chat_history",
+    "update_user_chat_history",
+    "get_user_response_language",
+    "set_user_response_language",
+    "get_user_agentic_strength",
+    "set_user_agentic_strength",
+    "set_first_interaction",
+    "is_first_interaction",
+]
