+++ tests/test_chroma_endpoints.py
@@ -5,7 +5,7 @@ import chromadb
 from fastapi.testclient import TestClient
 
 import bt_servant as api
-import db.chroma_db as cdb
+import bt_servant_engine.adapters.chroma as cdb
 from bt_servant_engine.core.config import config as app_config
